feat edit chart route
