+++ src/routes/revforce/dashboard/editchart/$chartId.tsx
@@ -1,0 +1,16 @@
+import { createFileRoute } from "@tanstack/react-router";
+import { ConfigureChart } from "@/components/ConfigureChart";
+
+export const Route = createFileRoute("/revforce/dashboard/editchart/$chartId")({
+  component: RouteComponent,
+});
+
+function RouteComponent() {
+  const { chartId } = Route.useParams();
+  
+  return (
+    <ConfigureChart
+      chartId={chartId}
+    />
+)
+}
